requirements.txt
pillow>=10.0.0
openai>=1.0.0
numpy>=1.24.0


